float instead double

+++ Source/main.c
@@ -1,6 +1,8 @@
 #include<stdio.h>
 int main() {
-  double first, second, temp;
+
+
+  float first, second, temp;
   printf("Enter first number: ");
   scanf("%lf", &first);
   printf("Enter second number: ");
@@ -18,6 +20,7 @@ int main() {
   // %.2lf displays number up to 2 decimal points
   printf("\nAfter swapping, first number = %.2lf\n", first);
   printf("After swapping, second number = %.2lf", second);
+
   return 0;
 }
